backend/requirements.txt
asgiref==3.2.10
Django==2.2.19
django-filter==2.4.0
django-cors-headers==3.9.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.2.1
gunicorn==20.0.4
Pillow==8.2.0
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
zipp==3.15.0
